requirements.txt
paramiko
pywinrm[credssp]
pypsexec
pyyaml
mac-vendor-lookup
pysnmp-lextudio